Improve WebSocket performance
