+++ src/net/rpc.rs
@@ -25,19 +25,25 @@ async fn handler(
 	Extension(sess): Extension<Session>,
 	Extension(req_id): Extension<RequestId>,
 ) -> impl IntoResponse {
-	// finalize the upgrade process by returning upgrade callback.
-	// we can customize the callback by sending additional info such as address.
-	ws.on_upgrade(move |socket| handle_socket(socket, sess, req_id))
+	ws
+		// Set the maximum frame size to 16mb
+		.max_frame_size(64 << 20)
+		// Set the maximum message size to 256mb
+		.max_message_size(256 << 20)
+		// Set the potential WebSocket protocol formats
+		.protocols(["surrealql-binary", "json", "cbor", "messagepack"])
+		// Handle the WebSocket upgrade and process messages
+		.on_upgrade(move |socket| handle_socket(socket, sess, req_id))
 }
 
 async fn handle_socket(ws: WebSocket, sess: Session, req_id: RequestId) {
+	// Create a new connection instance
 	let rpc = Connection::new(sess);
-
 	// Update the WebSocket ID with the Request ID
 	if let Ok(Ok(req_id)) = req_id.header_value().to_str().map(Uuid::parse_str) {
 		// If the ID couldn't be updated, ignore the error and keep the default ID
 		let _ = rpc.write().await.update_ws_id(req_id).await;
 	}
-
+	// Serve the socket connection requests
 	Connection::serve(rpc, ws).await;
 }
