+++ src/rpc/response.rs
@@ -1,3 +1,5 @@
+use crate::err;
+use crate::telemetry::metrics::ws::record_rpc;
 use axum::extract::ws::Message;
 use opentelemetry::Context as TelemetryContext;
 use serde::Serialize;
@@ -9,10 +11,6 @@ use surrealdb::dbs::Notification;
 use surrealdb::sql;
 use surrealdb::sql::Value;
 use tracing::Span;
-
-use crate::err;
-use crate::rpc::CONN_CLOSED_ERR;
-use crate::telemetry::metrics::ws::record_rpc;
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
 pub enum OutputFormat {
@@ -104,8 +102,8 @@ impl Response {
 				"otel.status_message",
 				format!("code: {}, message: {}", err.code, err.message),
 			);
-			span.record("rpc.jsonrpc.error_code", err.code);
-			span.record("rpc.jsonrpc.error_message", err.message.as_ref());
+			span.record("rpc.error_code", err.code);
+			span.record("rpc.error_message", err.message.as_ref());
 		}
 
 		let (res_size, message) = match out {
@@ -127,14 +125,9 @@ impl Response {
 			}
 		};
 
-		if let Err(err) = chn.send(message).await {
-			if err.to_string() != CONN_CLOSED_ERR {
-				error!("Error sending response: {}", err);
-				return;
-			}
+		if chn.send(message).await.is_ok() {
+			record_rpc(&TelemetryContext::current(), res_size, is_error);
 		};
-
-		record_rpc(&TelemetryContext::current(), res_size, is_error);
 	}
 }
