+++ tests/ws_integration.rs
@@ -1451,4 +1451,48 @@ mod ws_integration {
 		assert!(res.starts_with("surrealdb-"), "result: {}", res);
 		Ok(())
 	}
+
+	// Validate that the WebSocket is able to process multiple queries concurrently
+	#[test(tokio::test)]
+	async fn concurrency() -> Result<(), Box<dyn std::error::Error>> {
+		let (addr, _server) = common::start_server_with_defaults().await.unwrap();
+		let socket = &mut common::connect_ws(&addr).await?;
+
+		//
+		// Prepare the connection
+		//
+		let res = common::ws_signin(socket, USER, PASS, None, None, None).await;
+		assert!(res.is_ok(), "result: {:?}", res);
+		let res = common::ws_use(socket, Some("N"), Some("D")).await;
+		assert!(res.is_ok(), "result: {:?}", res);
+
+		//
+		// Run 5 queries that do a SLEEP and verify they all complete concurrently
+		//
+
+		// Send queries
+		for i in 0..5 {
+			let query = format!("SLEEP 1s; RETURN 'done-{}';", i);
+			let query_msg = json!({
+					"id": "1",
+					"method": "query",
+					"params": [query],
+			});
+			let res = common::ws_send_msg(socket, serde_json::to_string(&query_msg).unwrap()).await;
+			assert!(res.is_ok(), "result: {:?}", res);
+		}
+
+		// Wait for queries to complete and verify they all completed within 2 seconds (assume they are executed concurrently)
+		let msgs = common::ws_recv_all_msgs(socket, 5, Duration::from_secs(2)).await;
+		assert!(msgs.is_ok(), "Error waiting for messages: {:?}", msgs.err());
+
+		let msgs = msgs.unwrap();
+		assert!(
+			msgs.iter().all(|v| v["error"].is_null()),
+			"Unexpected error received: {:#?}",
+			msgs
+		);
+
+		Ok(())
+	}
 }
